Development

+++ nocode-itdevice/database/migrations/2023_10_01_110737_add_percent_to_attribute_values_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('attribute_values', function (Blueprint $table) {
-            $table->unsignedTinyInteger('percent');
+            $table->float('percent');
         });
     }
